requirements.txt
python-dotenv==0.15.0
requests==2.25.0
beautifulsoup4==4.9.3
pandas==1.1.5
tweepy==3.9.0
numpy~=1.19.3